+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Create Avatar component
+- Create profile Top Bar
+
 ## [0.46.2] - 2022-06-01
 
 ### Fixed